move fixtures into conftest
